requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
shim~=0.1.0
langchain-community~=0.4.1
langchain-text-splitters~=1.0.0
qdrant-client~=1.15.1
dspy~=3.0.3
fastapi~=0.120.0
pydantic~=2.12.3
uvicorn~=0.38.0
sentence-transformers==5.1.2
faiss-cpu


